Implement support for icon packs
It is now possible to use icon packs in S Launcher. You can select them
in the Launcher Settings using SharedPreferences. This commit does not
implement back images or mask images. This will come when support for
Adaptive Icons are added, but that will be a ways off.

To support icon packs, the following changes were made to existing code:

 - Added IconHandler and IconPackParser
 - ICON_PACK added to Preferences
 - SettingsActivity uses IconHandler to get list of icon packs
 - PreferenceUpdater is initialized in LauncherApplication
 - `loadApps(...)` has been moved out of ApplicationsModel